test(e2e): refactor code to improve reusability
Exported common functions for reuse across tests
Increased interval for a few more timeouts
